+++ new/src/AcademiiSdk/Model/CourseAssignUsersResponseInputData.cs
@@ -80,13 +80,13 @@ namespace AcademiiSdk.Model
         {
             var errors = new List<string>();
             // Assigned (long) maximum
-            if (this.Assigned > (long)9007199254740991)
+            if (this.Assigned > 9007199254740991)
             {
                 errors.Add("Invalid value for Assigned, must be a value less than or equal to 9007199254740991.");
             }
 
             // Assigned (long) minimum
-            if (this.Assigned < (long)0)
+            if (this.Assigned < 0)
             {
                 errors.Add("Invalid value for Assigned, must be a value greater than or equal to 0.");
             }
