+++ new/src/AcademiiSdk/Model/OrganizationGroupAddUserResponseInputData.cs
@@ -112,13 +112,13 @@ namespace AcademiiSdk.Model
                 }
 
             // AddedMembershipCount (long) maximum
-            if (this.AddedMembershipCount > (long)9007199254740991)
+            if (this.AddedMembershipCount > 9007199254740991)
             {
                 errors.Add("Invalid value for AddedMembershipCount, must be a value less than or equal to 9007199254740991.");
             }
 
             // AddedMembershipCount (long) minimum
-            if (this.AddedMembershipCount < (long)0)
+            if (this.AddedMembershipCount < 0)
             {
                 errors.Add("Invalid value for AddedMembershipCount, must be a value greater than or equal to 0.");
             }
